+++ test/Serilog.Sinks.HttpTests/Sinks/Http/Private/TextFormatters/CompactTextFormatterTest.cs
@@ -1,23 +1,23 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
 using Serilog.Events;
 using Serilog.Formatting;
 using Serilog.Sinks.Http.LogServer.Controllers.Dto;
+using Serilog.Sinks.Http.TextFormatters;
 using Serilog.Support;
 using Shouldly;
 using Xunit;
 
-namespace Serilog.Sinks.Http.Private.Formatters
+namespace Serilog.Sinks.Http.Private.TextFormatters
 {
-    public class NormalJsonFormatterTest
+    public class CompactTextFormatterTest
     {
         private readonly StringWriter output;
 
         private ILogger logger;
 
-        public NormalJsonFormatterTest()
+        public CompactTextFormatterTest()
         {
             output = new StringWriter();
         }
@@ -32,14 +32,22 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void LogEventLevels(LogEventLevel level)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(true));
+            logger = CreateLogger(new CompactRenderedTextFormatter());
 
             // Act
             logger.Write(level, "No properties");
 
             // Assert
             var @event = GetEvent();
-            @event.Level.ShouldNotBeNull();
+
+            if (level == LogEventLevel.Information)
+            {
+                @event.Level.ShouldBeNull();
+            }
+            else
+            {
+                @event.Level.ShouldNotBeNull();
+            }
         }
 
         [Theory]
@@ -48,7 +56,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void EmptyEvent(bool isRenderingMessage)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(isRenderingMessage));
+            logger = CreateLogger(isRenderingMessage ?
+                new CompactRenderedTextFormatter() :
+                new CompactTextFormatter());
 
             // Act
             logger.Information("No properties");
@@ -56,11 +66,10 @@ namespace Serilog.Sinks.Http.Private.Formatters
             // Assert
             var @event = GetEvent();
             @event.Timestamp.ShouldNotBeNull();
-            @event.Level.ShouldBe("Information");
+            @event.Level.ShouldBeNull();
             @event.MessageTemplate.ShouldBe("No properties");
             @event.RenderedMessage.ShouldBe(isRenderingMessage ? "No properties" : null);
             @event.Exception.ShouldBeNull();
-            @event.Properties.ShouldBeNull();
             @event.Renderings.ShouldBeNull();
         }
 
@@ -70,7 +79,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void MinimalEvent(bool isRenderingMessage)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(isRenderingMessage));
+            logger = CreateLogger(isRenderingMessage ?
+                new CompactRenderedTextFormatter() :
+                new CompactTextFormatter());
 
             // Act
             logger.Information("One {Property}", 42);
@@ -78,11 +89,10 @@ namespace Serilog.Sinks.Http.Private.Formatters
             // Assert
             var @event = GetEvent();
             @event.Timestamp.ShouldNotBeNull();
-            @event.Level.ShouldBe("Information");
             @event.MessageTemplate.ShouldBe("One {Property}");
             @event.RenderedMessage.ShouldBe(isRenderingMessage ? "One 42" : null);
             @event.Exception.ShouldBeNull();
-            @event.Properties.ShouldBe(new Dictionary<string, string> { { "Property", "42" } });
+            GetProperty("Property").ShouldBe("42");
             @event.Renderings.ShouldBeNull();
         }
 
@@ -92,7 +102,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void MultipleProperties(bool isRenderingMessage)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(isRenderingMessage));
+            logger = CreateLogger(isRenderingMessage ?
+                new CompactRenderedTextFormatter() :
+                new CompactTextFormatter());
 
             // Act
             logger.Information("Property {First} and {Second}", "One", "Two");
@@ -100,11 +112,11 @@ namespace Serilog.Sinks.Http.Private.Formatters
             // Assert
             var @event = GetEvent();
             @event.Timestamp.ShouldNotBeNull();
-            @event.Level.ShouldBe("Information");
             @event.MessageTemplate.ShouldBe("Property {First} and {Second}");
             @event.RenderedMessage.ShouldBe(isRenderingMessage ? "Property \"One\" and \"Two\"" : null);
             @event.Exception.ShouldBeNull();
-            @event.Properties.ShouldBe(new Dictionary<string, string> { { "First", "One" }, { "Second", "Two" } });
+            GetProperty("First").ShouldBe("One");
+            GetProperty("Second").ShouldBe("Two");
             @event.Renderings.ShouldBeNull();
         }
 
@@ -114,7 +126,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void Exceptions(bool isRenderingMessage)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(isRenderingMessage));
+            logger = CreateLogger(isRenderingMessage ?
+                new CompactRenderedTextFormatter() :
+                new CompactTextFormatter());
 
             // Act
             logger.Information(new DivideByZeroException(), "With exception");
@@ -122,11 +136,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
             // Assert
             var @event = GetEvent();
             @event.Timestamp.ShouldNotBeNull();
-            @event.Level.ShouldBe("Information");
             @event.MessageTemplate.ShouldBe("With exception");
             @event.RenderedMessage.ShouldBe(isRenderingMessage ? "With exception" : null);
             @event.Exception.ShouldNotBeNull();
-            @event.Properties.ShouldBeNull();
             @event.Renderings.ShouldBeNull();
         }
 
@@ -136,7 +148,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void ExceptionAndProperties(bool isRenderingMessage)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(isRenderingMessage));
+            logger = CreateLogger(isRenderingMessage ?
+                new CompactRenderedTextFormatter() :
+                new CompactTextFormatter());
 
             // Act
             logger.Information(new DivideByZeroException(), "With exception and {Property}", 42);
@@ -144,11 +158,10 @@ namespace Serilog.Sinks.Http.Private.Formatters
             // Assert
             var @event = GetEvent();
             @event.Timestamp.ShouldNotBeNull();
-            @event.Level.ShouldBe("Information");
             @event.MessageTemplate.ShouldBe("With exception and {Property}");
             @event.RenderedMessage.ShouldBe(isRenderingMessage ? "With exception and 42" : null);
             @event.Exception.ShouldNotBeNull();
-            @event.Properties.ShouldBe(new Dictionary<string, string> { { "Property", "42" } });
+            GetProperty("Property").ShouldBe("42");
             @event.Renderings.ShouldBeNull();
         }
 
@@ -158,7 +171,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void Renderings(bool isRenderingMessage)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(isRenderingMessage));
+            logger = CreateLogger(isRenderingMessage ?
+                new CompactRenderedTextFormatter() :
+                new CompactTextFormatter());
 
             // Act
             logger.Information("One {Rendering:x8}", 42);
@@ -166,14 +181,11 @@ namespace Serilog.Sinks.Http.Private.Formatters
             // Assert
             var @event = GetEvent();
             @event.Timestamp.ShouldNotBeNull();
-            @event.Level.ShouldBe("Information");
             @event.MessageTemplate.ShouldBe("One {Rendering:x8}");
             @event.RenderedMessage.ShouldBe(isRenderingMessage ? "One 0000002a" : null);
             @event.Exception.ShouldBeNull();
-            @event.Properties.ShouldBe(new Dictionary<string, string> { { "Rendering", "42" } });
-            @event.Renderings.ShouldContainKey("Rendering");
-            @event.Renderings["Rendering"].Length.ShouldBe(1);
-            @event.Renderings["Rendering"][0].ShouldBe(new RenderingDto { Format = "x8", Rendering = "0000002a" });
+            GetProperty("Rendering").ShouldBe("42");
+            @event.Renderings.ShouldBe(new[] { "0000002a" });
         }
 
         [Theory]
@@ -182,7 +194,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void MultipleRenderings(bool isRenderingMessage)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(isRenderingMessage));
+            logger = CreateLogger(isRenderingMessage ?
+                new CompactRenderedTextFormatter() :
+                new CompactTextFormatter());
 
             // Act
             logger.Information("Rendering {First:x8} and {Second:x8}", 1, 2);
@@ -190,17 +204,12 @@ namespace Serilog.Sinks.Http.Private.Formatters
             // Assert
             var @event = GetEvent();
             @event.Timestamp.ShouldNotBeNull();
-            @event.Level.ShouldBe("Information");
             @event.MessageTemplate.ShouldBe("Rendering {First:x8} and {Second:x8}");
             @event.RenderedMessage.ShouldBe(isRenderingMessage ? "Rendering 00000001 and 00000002" : null);
             @event.Exception.ShouldBeNull();
-            @event.Properties.ShouldBe(new Dictionary<string, string> { { "First", "1" }, { "Second", "2" } });
-            @event.Renderings.ShouldContainKey("First");
-            @event.Renderings["First"].Length.ShouldBe(1);
-            @event.Renderings["First"][0].ShouldBe(new RenderingDto { Format = "x8", Rendering = "00000001" });
-            @event.Renderings.ShouldContainKey("Second");
-            @event.Renderings["Second"].Length.ShouldBe(1);
-            @event.Renderings["Second"][0].ShouldBe(new RenderingDto { Format = "x8", Rendering = "00000002" });
+            GetProperty("First").ShouldBe("1");
+            GetProperty("Second").ShouldBe("2");
+            @event.Renderings.ShouldBe(new[] { "00000001", "00000002" });
         }
 
         [Theory]
@@ -209,7 +218,9 @@ namespace Serilog.Sinks.Http.Private.Formatters
         public void NastyException(bool isRenderingMessage)
         {
             // Arrange
-            logger = CreateLogger(new NormalJsonFormatter(isRenderingMessage));
+            logger = CreateLogger(isRenderingMessage ?
+                new CompactRenderedTextFormatter() :
+                new CompactTextFormatter());
 
             // Act
             logger.Information(new NastyException(), "With exception");
@@ -226,9 +237,15 @@ namespace Serilog.Sinks.Http.Private.Formatters
                 .CreateLogger();
         }
 
-        private EventDto GetEvent()
+        private CompactEventDto GetEvent()
         {
-            return JsonConvert.DeserializeObject<EventDto>(output.ToString());
+            return JsonConvert.DeserializeObject<CompactEventDto>(output.ToString());
+        }
+
+        private string GetProperty(string name)
+        {
+            dynamic @event = JsonConvert.DeserializeObject(output.ToString());
+            return @event[name];
         }
     }
 }